+++ app/controllers/conversions/involuntary/projects_controller.rb
@@ -43,8 +43,7 @@ class Conversions::Involuntary::ProjectsController < Conversions::ProjectsContro
       :establishment_sharepoint_link,
       :trust_sharepoint_link,
       :note_body,
-      :directive_academy_order,
-      :sponsor_trust_required
+      :directive_academy_order
     )
   end
 end