+++ db/migrate/20230404144130_remove_sponsor_trust_required.rb
@@ -1,0 +1,5 @@
+class RemoveSponsorTrustRequired < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :projects, :sponsor_trust_required, :boolean
+  end
+end