+++ spec/features/users_can_create_involuntary_conversion_projects_spec.rb
@@ -43,10 +43,6 @@ RSpec.feature "Users can create new involuntary conversion projects" do
         choose "Yes"
       end
 
-      within("#sponsor-trust-required") do
-        choose "Yes"
-      end
-
       click_button("Continue")
 
       expect(page).to have_content(I18n.t("project.show.title"))