+++ spec/features/users_can_create_voluntary_conversion_projects_spec.rb
@@ -103,8 +103,5 @@ RSpec.feature "Users can create new voluntary conversion projects" do
     within("#directive-academy-order") do
       choose "No"
     end
-    within("#sponsor-trust-required") do
-      choose "No"
-    end
   end
 end